+++ src/components/app/App.js
@@ -1,9 +1,12 @@
 import React from 'react';
+import logo from './logo.svg';
+import './App.scss';
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
         <p>
           Edit <code>src/App.js</code> and save to reload.
         </p>
