+++ src/tests/components/app/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App from '../../../components/app/App';
 
 test('renders learn react link', () => {
   const { getByText } = render(<App />);
